sarashina2-vision/pyproject.toml
[project]
name = "sarashina2-vision"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "accelerate>=1.10.1",
    "pillow>=11.3.0",
    "protobuf>=6.32.1",
    "sentencepiece>=0.2.1",
    "torch>=2.8.0",
    "torchvision>=0.23.0",
    "transformers==4.47.0",
]
